# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentDialog Agent API",
    "version": "dev",
    "description": "The surface an AI agent integrates against: conversations, messages, human queries and webhooks."
  },
  "servers": [
    {
      "url": "https://api.agentdialog.io"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/agent/queries/{id}/cancel": {
      "post": {
        "tags": [
          "queries"
        ],
        "summary": "Withdraw a query",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The query, cancelled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "query_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "assigned",
                            "needs_context",
                            "answered",
                            "expired",
                            "cancelled"
                          ]
                        },
                        "status_description": {
                          "type": "string"
                        },
                        "query_type": {
                          "type": "string",
                          "enum": [
                            "validation",
                            "interpretation",
                            "expert_query",
                            "labeling"
                          ]
                        },
                        "question": {
                          "type": "string"
                        },
                        "context": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "confidence": {
                          "type": [
                            "number",
                            "null"
                          ]
                        },
                        "language": {
                          "type": "string",
                          "enum": [
                            "en",
                            "es",
                            "ca"
                          ]
                        },
                        "answer": {
                          "type": [
                            "object",
                            "null"
                          ],
                          "additionalProperties": {}
                        },
                        "comment": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "human_confidence": {
                          "type": [
                            "number",
                            "null"
                          ]
                        },
                        "response_time_ms": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "insufficient_reason": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "expires_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "query_id",
                        "status",
                        "status_description",
                        "query_type",
                        "question",
                        "context",
                        "confidence",
                        "language",
                        "answer",
                        "comment",
                        "human_confidence",
                        "response_time_ms",
                        "insufficient_reason",
                        "created_at",
                        "expires_at"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "No such query, or it wasn't created by the authenticated agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The query already reached a terminal state and cannot be cancelled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An agent API key, prefixed `mge_ag_`."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "NOT_FOUND",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "CONFLICT",
                  "VALIDATION_ERROR",
                  "RATE_LIMIT",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "IDEMPOTENCY_KEY_REUSED",
                  "UNDECIDABLE_QUERY",
                  "PAYLOAD_TOO_LARGE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              },
              "retryAfter": {
                "type": "integer"
              },
              "reason": {
                "type": "string"
              },
              "detail": {
                "type": "string"
              },
              "remedy": {
                "type": "string"
              },
              "prior_query_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "The bearer token is missing, malformed, or does not match a live agent.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests. `retryAfter` in the error envelope is the number of seconds to wait before trying again.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An unexpected error. The error envelope carries no further detail than `message`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  }
}